Make the organization of the body and the summary have parallel structure.
This means I moved the 'new expected features' section to the top of the
summary, and made a new section with the same name at the start of
the body, turning the previous top level sections there into subsections.

I also added a line to the new first summary section for modulespec (pep 451).