whatsnew: unittest discover works on namespace packages (#17457).
I also added the docs that never got committed, editing the patch provided by
Claudiu.

I collapsed the now three versionchanged directives for discovery into one.  I
tried several different styles for this.  The most obvious is a bulleted list,
but currently the only way I could find to produce that is to have *something*
appear on the versionchanged line after the colon, which combined with the
blank space before the bulleted list just looks wrong.  I also tried the
current single-line-three-sentences version with bullet characters before each
sentence, and I almost went with that as it made it clear that the three
sentences are independent.  But I decided to just go with the sentences for
simplicity reasons.