gh-99942: python.pc on android/cygwin should link to libpython per configure.ac (GH-100356)
In commit 0cf5698cf573eb0922b9faff8dd7cae0e1065920 a previous change to avoid linking to libpython was partially reverted for Android (and later Cygwin as well), to add back the link flags. This was applied to distutils and to python-config.sh, but not to python.pc.

Add it back to python.pc as well.

Automerge-Triggered-By: GH:gpshead